backend/requirements.txt
asgiref==3.7.1
backports.zoneinfo==0.2.1
Django==4.2.2
django-filter==23.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==18.10.1
importlib-metadata==6.6.0
Markdown==3.4.3
mysqlclient==2.1.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.6.1
zipp==3.15.0
